Refactoring changes after meeting
